yolov5-prune/requirements.txt
# YOLOv5 🚀 requirements
# Usage: pip install -r requirements.txt -i https://mirrors.huaweicloud.com/repository/pypi/simple

# Base ------------------------------------------------------------------------
# based on python==3.8.17
ipython==8.10.0          # ! interactive notebook  if python3.8 need ipython at 8.1~8.12; latest version 8.13 need python3.9;
matplotlib==3.7.2
numpy==1.24.3            # !
opencv-python==4.8.0.76
Pillow==9.5              # !
psutil                   # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy==1.8.0             # !
# thop>=0.1.1              # FLOPs computation
# torch>=1.13.0             # see https://pytorch.org/get-started/locally (recommended)
# torchvision>=0.8.1 
tqdm>=4.64.0 
# protobuf<=3.20.1       # https://github.com/ultralytics/yolov5/issues/8012

# Logging ---------------------------------------------------------------------
tensorboard==2.13.0      # !
# clearml>=1.2.0    
# comet

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export ----------------------------------------------------------------------
coremltools>=6.0         # CoreML export
onnx==1.14.1             # ONNX export
onnx-simplifier==0.4.36  # ONNX simplifier
# nvidia-pyindex         # TensorRT export
# nvidia-tensorrt        # TensorRT export
scikit-learn==1.1.2      # ! CoreML quantization
tensorflow==2.13.0       # ! TF exports (-cpu, -aarch64, -macos)
tensorflowjs>=3.9.0      # TF.js export
openvino-dev             # OpenVINO export

# Deploy ----------------------------------------------------------------------
# tritonclient[all]~=2.24.0   # ? install problem

# Extras ----------------------------------------------------------------------
mss                      # screenshots
albumentations>=1.0.3
pycocotools>=2.0         # COCO mAP
roboflow
ultralytics              # HUB https://hub.ultralytics.com
